Use the stock camera interface blob from stock
Why? Because at the moment we cannot figure out why the source built version is failing to probe the sensors.
In theory, the stock one should be fine as a brief examination of the stock one showed that there were no obvious changes by lg.
It also contains more or less identical functions. So, for the time being, let's use the stock one for now.